backend/project/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.48.0
idna==3.11
pillow==12.1.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
rsa==4.9.1
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
tzdata==2025.3
urllib3==2.6.3
